+++ arrow/src/compute/kernels/mod.rs
@@ -25,13 +25,13 @@ pub mod cast;
 pub mod cast_utils;
 pub mod comparison;
 pub mod concat;
+pub mod concat_elements;
 pub mod filter;
 pub mod length;
 pub mod limit;
 pub mod partition;
 pub mod regexp;
 pub mod sort;
-pub mod string;
 pub mod substring;
 pub mod take;
 pub mod temporal;
